refactor(formats): split gene-model gff parser internals
