+++ tests/test_gene_model_gff_parser.py
@@ -4,10 +4,13 @@ import inspect
 
 from SpliceGrapher.formats import gene_model as gm
 from SpliceGrapher.formats.parsers import gene_model_gff as parser_boundary
+from SpliceGrapher.formats.parsers import gene_model_gff_context as parser_context
+from SpliceGrapher.formats.parsers import gene_model_gff_handlers as parser_handlers
+from SpliceGrapher.formats.parsers import gene_model_gff_records as parser_records
 
 
 def test_candidate_cache_uses_bounded_clear_policy() -> None:
-    ctx = parser_boundary.ParseContext(
+    ctx = parser_context.ParseContext(
         model=gm.GeneModel(),
         require_notes=False,
         verbose=False,
@@ -15,9 +18,9 @@ def test_candidate_cache_uses_bounded_clear_policy() -> None:
         parent_candidate_max_entries=2,
     )
 
-    parser_boundary._candidate_parent_ids(ctx, "GENE1.1")
-    parser_boundary._candidate_parent_ids(ctx, "GENE2.1")
-    parser_boundary._candidate_parent_ids(ctx, "GENE3.1")
+    parser_handlers.candidate_parent_ids(ctx, "GENE1.1")
+    parser_handlers.candidate_parent_ids(ctx, "GENE2.1")
+    parser_handlers.candidate_parent_ids(ctx, "GENE3.1")
 
     assert len(ctx.parent_candidates) <= 2
     assert ctx.stats.parent_candidate_clear_count == 1
@@ -31,7 +34,7 @@ def test_parent_cache_uses_bounded_clear_policy() -> None:
     model.add_gene(gene_one)
     model.add_gene(gene_two)
 
-    ctx = parser_boundary.ParseContext(
+    ctx = parser_context.ParseContext(
         model=model,
         require_notes=False,
         verbose=False,
@@ -39,15 +42,22 @@ def test_parent_cache_uses_bounded_clear_policy() -> None:
         parent_cache_max_entries=1,
     )
 
-    assert parser_boundary._resolve_parent(ctx, "GENE1", "chr1") is gene_one
-    assert parser_boundary._resolve_parent(ctx, "GENE2", "chr1") is gene_two
+    assert parser_handlers.resolve_parent(ctx, "GENE1", "chr1") is gene_one
+    assert parser_handlers.resolve_parent(ctx, "GENE2", "chr1") is gene_two
 
     assert len(ctx.parent_cache) == 1
     assert ctx.stats.parent_cache_clear_count == 1
 
 
-def test_parser_boundary_depends_on_models_module_not_gene_model_facade() -> None:
-    source = inspect.getsource(parser_boundary)
+def test_parser_modules_depend_on_models_module_not_gene_model_facade() -> None:
+    source = "\n".join(
+        [
+            inspect.getsource(parser_boundary),
+            inspect.getsource(parser_context),
+            inspect.getsource(parser_records),
+            inspect.getsource(parser_handlers),
+        ]
+    )
 
     assert "import SpliceGrapher.formats.models as model_domain" in source
     assert "import SpliceGrapher.formats.gene_model as gm" not in source
